clase5-relaciones/requirements.txt
flask
python-dotenv
flask-restful
flask_sqlalchemy
